ocr/requirements.txt
# Dependencies for the OCR container.

fastapi==0.115.12
uvicorn[standard]==0.34.2
pytesseract
Pillow
fastapi
uvicorn
